renamed generic_storage_pyo file and test file
