+++ h2integrate/control/test/test_openloop_controllers.py
@@ -7,7 +7,7 @@ import openmdao.api as om
 from pytest import fixture
 
 from h2integrate.core.file_utils import load_yaml
-from h2integrate.storage.generic_storage_pyo import StoragePerformanceModel
+from h2integrate.storage.storage_performance_model import StoragePerformanceModel
 from h2integrate.control.control_strategies.storage.simple_openloop_controller import (
     SimpleStorageOpenLoopController,
 )
