+++ h2integrate/control/test/test_openloop_storage_integration.py
@@ -3,8 +3,8 @@ import pytest
 import openmdao.api as om
 from pytest import fixture
 
-from h2integrate.storage.generic_storage_pyo import StoragePerformanceModel
 from h2integrate.storage.battery.pysam_battery import PySAMBatteryPerformanceModel
+from h2integrate.storage.storage_performance_model import StoragePerformanceModel
 from h2integrate.storage.simple_storage_auto_sizing import StorageAutoSizingModel
 from h2integrate.control.control_strategies.storage.simple_openloop_controller import (
     SimpleStorageOpenLoopController,
