+++ h2integrate/storage/test/test_storage_performance_model.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 import openmdao.api as om
 
-from h2integrate.storage.generic_storage_pyo import StoragePerformanceModel
+from h2integrate.storage.storage_performance_model import StoragePerformanceModel
 from h2integrate.control.control_strategies.storage.simple_openloop_controller import (
     SimpleStorageOpenLoopController,
 )
